Valid phone number
The program receives a line of text as input. Write a program that determines whether the entered string is a valid phone number. The text string is a valid phone number if it has the format:

abc-def-hijk or
7-abc-def-hijk
where a, b, c, d, e, f, h, i, j, k are digits from 0 to 9.